Refactoring and destructuring in hook
